Added Broadcaster
The Broadcaster is a class that implements a simple distribution
mechanism for events. It extends and implements both the event::Listener
and the event::Service interface. Basically, you can subscribe Listeners
to it that will later get the events the Broadcaster gets.
This is useful because most classes that provide a service would
probably implement it in about the same way, so the implementation can
be decoupled here.