Add setup-specific files to .gitignore
I sometimes use cscope to browse through my code (integrated with vim)
and I wrote a makefile that wraps the build setup to integrate with vim.
These are changes that should be kept on my computer only and not pushed
up to sites like github, since they are specific to my setup.