Improve example dependency tree
"events" might be misunderstood as the events that are sent around in
the game. Meant was a library similar to the one I made for my packman
game. (https://github.com/shak-mar/cpp-games/tree/master/library/events)
I instead use a "geometry" library for my example, which is probably
understood as more general.